+++ src/main/java/com/vincevscode/cointracker/repository/AuthUserRepositoryInterface.java
@@ -2,9 +2,12 @@
 package com.vincevscode.cointracker.repository;
 
 import com.vincevscode.cointracker.model.AuthUser;
+import com.vincevscode.cointracker.model.UserRole;
 
 public interface AuthUserRepositoryInterface {
     AuthUser findAuthUserById(int userId);
 
     AuthUser findAuthUserByUsername(String username);
+
+    AuthUser createAuthUser(String username, String passwordHash, UserRole role, boolean active);
 }